refactor(image-viewer): colocate all viewer props in viewerTypes
- Move ImageViewerProps into viewerTypes.ts alongside Display/Controls props
- Export ImageViewerProps from the types barrel, not the component file
- Keep a single place to read viewer public prop shapes
